Restructure repository layout into scripts/data/docs/config
