+++ data/oracle_ddl/california_schools/california_schools_metadata.sql
