+++ data/oracle_ddl/card_games/card_games_oracle.sql
