+++ data/oracle_ddl/codebase_community/codebase_community_oracle.sql
