+++ data/oracle_ddl/debit_card_specializing/debit_card_specializing_metadata.sql
