+++ data/oracle_ddl/european_football_2/european_football_2_metadata.sql
