+++ data/oracle_ddl/financial/financial_metadata.sql
