+++ data/oracle_ddl/financial/financial_oracle.sql
