+++ data/oracle_ddl/student_club/student_club_metadata.sql
