+++ data/oracle_ddl/thrombosis_prediction/thrombosis_prediction_oracle.sql
