+++ sql/index_creation.sql
